backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
yt-dlp
pdfplumber
sumy
numpy
youtube-transcript-api
google-generativeai
requests
aiofiles